Removing the framework_tools.py file
I ran all of our tests to make sure it still works, still going to split
the test file for framework_tools
