+++ modules/toolbox/validation_package.py
@@ -1,4 +1,5 @@
-from modules.toolbox import framework_tools as ft
+from modules.DMZ.data_kit import data_splitting
+from modules.DMZ.data_kit import data_io
 from modules.toolbox import splitter as sp
 import setup
 
@@ -12,12 +13,12 @@ class ValidationPackage(object):
         self.y_test = None
 
     def setup_package(self, pack):
-        self.x_train, self.x_test, self.y_train, self.y_test = ft.get_train_test(pack.train_data, pack.target_column)
+        self.x_train, self.x_test, self.y_train, self.y_test = data_splitting.get_train_test(pack.train_data, pack.target_column)
 
     def split_file(self, train_file, target):
         splitter = sp.Splitter()
         splitter.Split(setup.get_datasets_path() + train_file, target)
-        train = ft.get_data(setup.get_datasets_path(), "training.data.csv", ",")
-        test = ft.get_data(setup.get_datasets_path(), "testing.data.csv", ",")
-        self.x_train, self.y_train = ft.separate_target(train, target)
-        self.x_test, self.y_test = ft.separate_target(test, target)
+        train = data_io.get_data(setup.get_datasets_path(), "training.data.csv", ",")
+        test = data_io.get_data(setup.get_datasets_path(), "testing.data.csv", ",")
+        self.x_train, self.y_train = data_splitting.separate_target(train, target)
+        self.x_test, self.y_test = data_splitting.separate_target(test, target)
