+++ tests/test_component_analyzers.py
@@ -4,7 +4,7 @@ import sys
 sys.path.insert(0, os.path.abspath('..'))
 
 from unittest import TestCase
-from modules.toolbox import framework_tools as ft
+from modules.DMZ.data_kit import data_io
 from modules.reducers import scikit_component_analyzers
 from sklearn.random_projection import GaussianRandomProjection
 from sklearn.decomposition import PCA
@@ -24,7 +24,7 @@ class TestDimensionalityReduction(TestCase):
         end = analyzer.shape
 
         # Assert fewer rows
-        self.assertLess(end[0], start[0], msg = "Failed to beat baseline")
+        self.assertLess(end[0], start[0], msg="Failed to beat baseline")
 
     def test_independent_component_analyzer(self):
 
@@ -36,7 +36,7 @@ class TestDimensionalityReduction(TestCase):
         end = analyzer.shape
 
         # Assert fewer rows
-        self.assertLess(end[0], start[0], msg = "Failed to beat baseline")
+        self.assertLess(end[0], start[0], msg="Failed to beat baseline")
 
     def test_factor_component_analyzer(self):
 
@@ -63,7 +63,7 @@ class TestDimensionalityReduction(TestCase):
         self.assertLess(end[0], start[0], msg="Failed to beat baseline")
 
     def pre_test(self):
-        data = ft.get_data(setup.get_datasets_path(), "winequality-red.csv", ';')
+        data = data_io.get_data(setup.get_datasets_path(), "winequality-red.csv", ';')
         start = data.shape
 
         return data, start
